apps/agents/pyproject.toml
[project]
name = "aip-agents"
version = "0.1.0"
description = "FastAPI + Google ADK Agents Application"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.121.2",
    "uvicorn[standard]==0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "google-cloud-firestore>=2.18.0",
    "google-cloud-storage>=2.18.0",
    "google-cloud-logging>=3.11.0",
    "google-cloud-trace>=1.15.0",
    "httpx>=0.27.0",
    "structlog>=24.4.0",
    "sentry-sdk[fastapi]>=2.17.0",
    "python-dotenv>=1.0.1",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.2.0",
    "tenacity>=9.0.0",
    "slowapi>=0.1.9",
    "redis>=5.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
    "mypy>=1.11.0",
    "pre-commit>=3.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "DTZ"]
ignore = []
target-version = "py311"

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*" = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --cov=app --cov-report=html --cov-report=term-missing"

